Move StackActions to Stack
With StackConfigData now its own data class we've moved the methods that
were originally on Stack in v1 back to Stack.
